@@ -2,26 +2,27 @@
     import { API_URL } from '$lib/config';
     import { auth } from '$lib/stores/auth';
     import { createEventDispatcher } from 'svelte';
-    import Modal from './Modal.svelte';
+    import Modal from '../core/Modal.svelte';
+    import Button from '../core/Button.svelte';
 
     export let show = false;
+
     const dispatch = createEventDispatcher();
     let username = '';
     let password = '';
     let error = '';
 
-
-    async function handleSignup(event: SubmitEvent) {
+    async function handleLogin(event: SubmitEvent) {
         event.preventDefault();
         error = '';
 
         try {
-            const response = await fetch(`${API_URL}/auth/signup`, {
+            const response = await fetch(`${API_URL}/auth/login`, {
                 method: 'POST',
-                credentials: 'include',
                 headers: {
                     'Content-Type': 'application/json',
                 },
+                credentials: 'include',
                 body: JSON.stringify({ username, password }),
             });
 
@@ -29,21 +30,26 @@
                 const session = await response.json();
                 auth.setUser(session.user);
                 dispatch('close');
-                dispatch('success');
                 show = false;
             } else {
-                error = 'Signup failed';
+                error = 'Invalid credentials';
             }
         } catch (err) {
-            console.error('Signup error:', err);
-            error = 'Signup failed';
+            console.error('Login error:', err);
+            error = 'Login failed';
         }
+    }
+
+    function handleBackgroundClick() {
+        dispatch('close');
     }
 </script>
 
-<Modal bind:show={show} bind:error={error}>
-    <h2>Sign Up</h2>
-    <form on:submit={handleSignup}>
+<Modal bind:error bind:show closeButton>
+    <div slot="header">
+        <h2>Login</h2>
+    </div>
+    <form on:submit={handleLogin}>
         <div class="form-group">
             <label>Username:</label>
             <input type="text" bind:value={username} required>
@@ -52,37 +58,16 @@
             <label>Password:</label>
             <input type="password" bind:value={password} required>
         </div>
-        {#if error}
-            <div class="error">{error}</div>
-        {/if}
+
         <div class="buttons">
-            <button type="submit">Sign Up</button>
-            <button type="button" on:click={() => dispatch('close')}>Cancel</button>
+            <Button submit colour='blue'>Login</Button>
+            <Button onclick={() => dispatch('close')}>Cancel</Button>
         </div>
     </form>
+
 </Modal>
 
-
 <style>
-    .modal {
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        background-color: rgba(0,0,0,0.5);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .modal-content {
-        background-color: white;
-        padding: 20px;
-        width: 300px;
-        border-radius: 5px;
-    }
-
     .form-group {
         margin-bottom: 15px;
     }
@@ -91,11 +76,6 @@
         width: 100%;
         padding: 8px;
         margin-top: 5px;
-    }
-
-    .error {
-        color: red;
-        margin-bottom: 10px;
     }
 
     .buttons {
