@@ -3,8 +3,8 @@
     import { page } from '$app/stores';
     import { onMount } from 'svelte';
     import type { Character } from '$lib/types/types';
-    import LibrarySelector from '$lib/components/LibrarySelector.svelte';
-    import EncounterCreator from "$lib/components/EncounterCreator.svelte";
+    import LibrarySelector from '$lib/components/selectors/LibrarySelector.svelte';
+    import EncounterCreator from "$lib/components/encounter/EncounterCreator.svelte";
     import { id } from 'date-fns/locale';
     import { 
         getExperienceFromLevel, 
@@ -33,6 +33,10 @@
   import { creatureStore, hazardStore, itemStore } from '$lib/stores/libraryStore';
   import { characterStore } from '$lib/stores/characters';
   import { campaignSessionStore } from '$lib/stores/campaignSessions';
+    import Card from '$lib/components/core/Card.svelte';
+    import Button from '$lib/components/core/Button.svelte';
+    import ConfirmationModal from '$lib/components/modals/ConfirmationModal.svelte';
+    import Modal from '$lib/components/core/Modal.svelte';
 
 library.add(faLink)
 
@@ -42,7 +46,7 @@ library.add(faLink)
     let encounterCreator: EncounterCreator;
 
     // Variables for encounter display
-    let encountersListOpen = $state(true);
+    let encountersListClosed = $state(false);
     let encounterOpenStates: { [key: number]: boolean } = $state({});
     let encounterFilter = $state('');
     let encounterSort: 'name' | 'level' | 'xp' = $state('name');
@@ -237,7 +241,6 @@ library.add(faLink)
             encounterStore.unlinkEncounterFromSession(encounter.id);
         }
     }
-
 </script>
 
 <div class="encounters-page">
@@ -246,262 +249,220 @@ library.add(faLink)
         <div class="error">{error}</div>
     {/if}
 
+    <div class="creator">
     <EncounterCreator bind:editingEncounter bind:chosenSessionId bind:returnToSessionId bind:this={encounterCreator} />
-
+</div>
     {#if loading}
         <div class="loading">Loading encounters...</div>
     {:else}
-        <div class="encounters-section">
-            <div class="section-header" on:click={() => encountersListOpen = !encountersListOpen}>
+        <Card bind:collapsed={encountersListClosed}>
+            <div slot="header">
                 <h2>
                     Existing Encounters ({filteredAndSortedEncounters.length})
-                    <span class="toggle-icon">{encountersListOpen ? '▼' : '▶'}</span>
                 </h2>
             </div>
+                <div class="filter-sort">
+                    <input
+                        type="text"
+                        placeholder="Filter encounters..."
+                        bind:value={encounterFilter}
+                        class="filter-input"
+                    />
+                    <div class="sort-controls">
+                        <select bind:value={encounterSort}>
+                            <option value="name">Sort by Name</option>
+                            <option value="level">Sort by Level</option>
+                            <option value="xp">Sort by XP</option>
+                        </select>
+                        <Button colour='white' onclick={() => sortDirection = sortDirection === 'asc' ? 'desc' : 'asc'}>
+                           {sortDirection === 'asc' ? '↑' : '↓'}
 
-            {#if encountersListOpen}
-                <div class="encounters-controls" transition:fade>
-                    <div class="filter-sort">
-                        <input
-                            type="text"
-                            placeholder="Filter encounters..."
-                            bind:value={encounterFilter}
-                            class="filter-input"
-                        />
-                        <div class="sort-controls">
-                            <select bind:value={encounterSort}>
-                                <option value="name">Sort by Name</option>
-                                <option value="level">Sort by Level</option>
-                                <option value="xp">Sort by XP</option>
-                            </select>
-                            <button 
-                                class="sort-direction"
-                                on:click={() => sortDirection = sortDirection === 'asc' ? 'desc' : 'asc'}
-                            >
-                                {sortDirection === 'asc' ? '↑' : '↓'}
-                            </button>
-                        </div>
+                        </Button>
+   
                     </div>
                 </div>
-
-                <div class="encounters-list" transition:fade>
-                    {#each filteredAndSortedEncounters as encounter (encounter.id)}
-                        <div class="encounter-card">
-                            <div 
-                                class="encounter-header"
-                                on:click={() => encounterOpenStates[encounter.id] = !encounterOpenStates[encounter.id]}
-                            >
-                                <div class="encounter-summary">
-                                    <h3>{encounter.name}</h3>
-                                    <div class="encounter-meta">
-                                        <span class="status {encounter.status.toLowerCase()}">{encounter.status}</span>
-                                        <span class="xp">XP: {encounter.total_experience} (<span class="{getClassForDifficulty(getSeverityFromFinalExperience(encounter.total_experience, encounter.extra_experience))}">{getSeverityFromFinalExperience(encounter.total_experience, encounter.extra_experience).toWellFormed()}</span>)</span>
-                                        <span class="party">Level {encounter.party_level} ({encounter.party_size} players)</span>
-                                        {#if encounter.session_id}
-                                            <span class="session">Session: {sessionIx.get(encounter.session_id)}</span>
-                                        {/if}
-                                    </div>
-                                </div>
-                                <span class="toggle-icon">{encounterOpenStates[encounter.id] ? '▼' : '▶'}</span>
-                            </div>
-                            
-                            {#if encounterOpenStates[encounter.id]}
-                                <div class="encounter-details" transition:fade>
-                                    <p>{encounter.description}</p>
-                                    
-                                    <div class="details">
-                                        {#if encounter.enemies}
-                                        <div class="detail-section">
-                                            <h4>Enemies ({encounter.enemies.length})</h4>
-                                            <ul>
-                                                {#each encounter.enemies as encounterEnemy : EncounterEnemy}
-                                                    {#if getEnemyDetails(encounterEnemy.id)}
-                                                        <li>{getEnemyDetails(encounterEnemy.id)?.name} 
-                                                            {#if encounterEnemy.level_adjustment !== 0}
-                                                                ({getAdjustmentName(encounterEnemy.level_adjustment)})
-                                                            {/if}
-                                                            (Level {(getEnemyDetails(encounterEnemy.id)?.level || 0) + encounterEnemy.level_adjustment})
-                                                            (XP: {getExperienceFromLevel(encounter.party_level, getEnemyDetails(encounterEnemy.id)?.level || 0)})</li>
-                                                    {/if}
-                                                {/each}
-                                            </ul>
-                                        </div>
-                                        {/if}
-                                        {#if encounter.hazards}
-
-                                        <div class="detail-section">
-                                            <h4>Hazards ({encounter.hazards.length})</h4>
-                                            <ul>
-                                                {#each encounter.hazards as hazardId}
-                                                    {#if getHazardDetails(hazardId)}
-                                                        <li>{getHazardDetails(hazardId)?.name} (XP: {getExperienceFromLevel(encounter.party_level, getHazardDetails(hazardId)?.level || 0)})</li>
-                                                    {/if}
-                                                {/each}
-                                            </ul>
-                                        </div>
-                                        {/if}
-                                        {#if encounter.subsystem_type}
-                                        <div class="detail-section">
-                                            <h4>Subsystem</h4>
-                                            <p>Subsystem Type: {encounter.subsystem_type}</p>
-                                            <ul>
-                                                {#each encounter.subsystem_checks || [] as check}
-                                                    <li> {check.name} 
-
-                                                        ({#each check.roll_options as roll, i}
-                                                        {roll.skill} DC {roll.dc}{#if i < check.roll_options.length - 1},&nbsp;{/if} 
-                                                        {/each})
-
-                                                    </li>
-                                                    
-                                                    
-                                                {/each}
-                                            </ul>
-                                        </div>
-
-
-                                        {/if}
-
-                                        <div class="detail-section">
-                                            <h4>Treasure</h4>
-                                            <p>Currency: {encounter.treasure_currency}</p>
-                                            <ul>
-                                                {#each encounter.treasure_items as itemId}
-                                                    {#if getItemDetails(itemId)}
-                                                        <li>{getItemDetails(itemId)?.name}</li>
-                                                    {/if}
-                                                {/each}
-                                            </ul>
-                                        </div>
-                                    </div>
-                                    <div class="actions">
-
-                                    {#if encounter.status === 'Draft'}
-                                    <button 
-                                    class="clone-encounter-button"
-                                    on:click={() =>  encounterCreator.loadEncounterCopyToDraft(encounter)}
-                                >
-                                    Load draft
-                                </button>
-
-                                    {:else}
-                                    {#if !encounter.session_id}
-                                    <button 
-                                        class="complete-button"
-                                        disabled={encounter.status !== 'Prepared'}
-                                        on:click={() => linkingEncounter = encounter}
-                                    >
-                                        Link to session
-                                    </button>
-                                    {:else}
-                                    
-                                     
-                                    <button 
-                                        class="delete-button"
-                                        disabled={encounter.status !== 'Prepared'}
-                                        on:click={() => linkEncounterToSession(encounter, null)}
-                                    >
-                                        Unlink from session
-                                    </button>
-                                    {/if}
-
-                                            <button 
-                                                class="edit-button"
-                                                on:click={() => editingEncounter = encounter}
-                                            >
-                                                Edit
-                                            </button>
-                                            <button 
-                                            class="clone-encounter-button"
-                                            on:click={() =>  encounterCreator.loadEncounterCopyToDraft(encounter)}
-                                        >
-                                            Clone into draft
-                                        </button>
-
-                                        <button 
-                                        class="delete-button"
-                                        on:click={() => deletingEncounter = encounter.id}
-                                    >
-                                        Delete
-                                    </button>
-                                    {/if}
-
-
-                                        </div>
-                                        
-                                </div>
+            <Card background="light">
+                {#each filteredAndSortedEncounters as encounter (encounter.id)}
+                <Card bind:collapsed={
+                        () => encounterOpenStates[encounter.id]  ?? true,
+                        (val) => encounterOpenStates[encounter.id] = val}
+                    >
+                    <div slot="header" class="encounter-summary">
+                        <h3>{encounter.name}</h3>
+                        <div class="encounter-meta">
+                            <span class="status {encounter.status.toLowerCase()}">{encounter.status}</span>
+                            <span class="xp">XP: {encounter.total_experience} (<span class="{getClassForDifficulty(getSeverityFromFinalExperience(encounter.total_experience, encounter.extra_experience))}">{getSeverityFromFinalExperience(encounter.total_experience, encounter.extra_experience).toWellFormed()}</span>)</span>
+                            <span class="party">Level {encounter.party_level} ({encounter.party_size} players)</span>
+                            {#if encounter.session_id}
+                                <span class="session">Session: {sessionIx.get(encounter.session_id)}</span>
                             {/if}
                         </div>
-                    {/each}
-                </div>
-            {/if}
-        </div>
+                    </div>
+                    <!-- TODO: You have an encounter viewer modal, switch it out for this-->
+                    <div class="encounter-details">
+                        <p>{encounter.description}</p>
+                        
+                        <div class="details">
+                            {#if encounter.enemies}
+                            <div class="detail-section">
+                                <h4>Enemies ({encounter.enemies.length})</h4>
+                                <ul>
+                                    {#each encounter.enemies as encounterEnemy : EncounterEnemy}
+                                        {#if getEnemyDetails(encounterEnemy.id)}
+                                            <li>{getEnemyDetails(encounterEnemy.id)?.name} 
+                                                {#if encounterEnemy.level_adjustment !== 0}
+                                                    ({getAdjustmentName(encounterEnemy.level_adjustment)})
+                                                {/if}
+                                                (Level {(getEnemyDetails(encounterEnemy.id)?.level || 0) + encounterEnemy.level_adjustment})
+                                                (XP: {getExperienceFromLevel(encounter.party_level, getEnemyDetails(encounterEnemy.id)?.level || 0)})</li>
+                                        {/if}
+                                    {/each}
+                                </ul>
+                            </div>
+                            {/if}
+                            {#if encounter.hazards}
+
+                            <div class="detail-section">
+                                <h4>Hazards ({encounter.hazards.length})</h4>
+                                <ul>
+                                    {#each encounter.hazards as hazardId}
+                                        {#if getHazardDetails(hazardId)}
+                                            <li>{getHazardDetails(hazardId)?.name} (XP: {getExperienceFromLevel(encounter.party_level, getHazardDetails(hazardId)?.level || 0)})</li>
+                                        {/if}
+                                    {/each}
+                                </ul>
+                            </div>
+                            {/if}
+                            {#if encounter.subsystem_type}
+                            <div class="detail-section">
+                                <h4>Subsystem</h4>
+                                <p>Subsystem Type: {encounter.subsystem_type}</p>
+                                <ul>
+                                    {#each encounter.subsystem_checks || [] as check}
+                                        <li> {check.name} 
+
+                                            ({#each check.roll_options as roll, i}
+                                            {roll.skill} DC {roll.dc}{#if i < check.roll_options.length - 1},&nbsp;{/if} 
+                                            {/each})
+
+                                        </li>
+                                        
+                                        
+                                    {/each}
+                                </ul>
+                            </div>
+
+
+                            {/if}
+
+                            <div class="detail-section">
+                                <h4>Treasure</h4>
+                                <p>Currency: {encounter.treasure_currency}</p>
+                                <ul>
+                                    {#each encounter.treasure_items as itemId}
+                                        {#if getItemDetails(itemId)}
+                                            <li>{getItemDetails(itemId)?.name}</li>
+                                        {/if}
+                                    {/each}
+                                </ul>
+                            </div>
+                        </div>
+                        <div class="actions">
+
+                        {#if encounter.status === 'Draft'}
+                        <Button colour='blue' onclick={() => encounterCreator.loadEncounterCopyToDraft(encounter)}>
+                            Load draft
+
+                        </Button>
+
+                        {:else}
+                        {#if !encounter.session_id}
+
+                        <Button colour='green' disabled={encounter.status !== 'Prepared'} onclick={() => linkingEncounter = encounter}>
+                            Link to session
+                        </Button>
+                        
+                        {:else}
+                        
+                
+                        <Button colour='red' disabled={encounter.status !== 'Prepared'} onclick={() => linkEncounterToSession(encounter, null)}>
+                            Unlink from session
+                        </Button>
+                        {/if}
+
+                                <Button colour='blue' onclick={() => editingEncounter = encounter}>
+                                    Edit
+                                </Button>
+                                
+                            <Button colour='blue' onclick={() =>  encounterCreator.loadEncounterCopyToDraft(encounter)}>
+                                Clone into draft
+                            </Button>
+
+                        <Button colour='red' onclick={() => deletingEncounter = encounter.id}>
+                            Delete
+                        </Button>
+                        {/if}
+
+
+                            </div>
+                            
+                    </div>
+
+                </Card>
+                {/each}
+            </Card>
+        </Card>
     {/if}
     
 </div>
 
-{#if deletingEncounter && deletingEncounterName}
-    <div class="modal">
-        <div class="modal-content">
-            <h2>Delete Encounter: {deletingEncounterName}</h2>
+<ConfirmationModal show={!!deletingEncounter && !!deletingEncounterName}
+ on:confirm={() => deleteEncounter(deletingEncounter)} on:close={() => deletingEncounter = null}
+ confirmationString="Delete"
+>
 
-            <p>Are you sure you want to delete this encounter?</p>
-            <p>It will be deleted from any attached sessions and campaigns.</p>
+    <p>Are you sure you want to delete this encounter?</p>
+    <p>It will be deleted from any attached sessions and campaigns.</p>
 
-            <div class="modal-actions">
-                <button 
-                    class="cancel-button"
-                    on:click={() => deletingEncounter = null}
-                >
-                    Cancel
-                </button>
-                <button 
-                    class="delete-button"
-                    on:click={() => deleteEncounter(deletingEncounter)}
-                >
-                    Delete
-                </button>
-            </div>
-        </div>
+</ConfirmationModal>
+
+<Modal show={!!linkingEncounter} closeButton on:close={() => linkingEncounter = null}>
+    <div slot="header">
+    <h2>Link Encounter to Session</h2>
     </div>
-{/if}
 
-{#if linkingEncounter}
-    <div class="modal">
-        <div class="modal-content">
-            <h2>Link Encounter to Session</h2>
+    <p>Choose a session to link this encounter to:</p>
 
-            <p>Choose a session to link this encounter to:</p>
+    <select bind:value={selectedLinkingSession}>
+        <option value={null}>Select a session...</option>
+        {#each campaignSessions as session, ix}
+            <option value={session.id}>Session {ix}: {session.name}</option>
+        {/each}
+    </select>
 
-            <select bind:value={selectedLinkingSession}>
-                <option value={null}>Select a session...</option>
-                {#each campaignSessions as session, ix}
-                    <option value={session.id}>Session {ix}: {session.name}</option>
-                {/each}
-            </select>
+    <div class="modal-actions">
 
-            <div class="modal-actions">
-                <button 
-                    class="cancel-button"
-                    on:click={() => linkingEncounter = null}
-                >
-                    Cancel
-                </button>
-                <button 
-                    class="complete-button"
-                    disabled={selectedLinkingSession === null}
-                    on:click={() => {
-                        if (linkingEncounter === null || selectedLinkingSession === null) return;
-                        linkEncounterToSession(linkingEncounter, selectedLinkingSession);
-                        linkingEncounter = null;
-                    }}
-                >
-                    Link
-                </button>
-            </div>
-        </div>
+        <Button 
+            onclick={() => linkingEncounter = null}
+        >
+            Cancel
+</Button>
+
+        <Button 
+            colour='green'
+            disabled={selectedLinkingSession === null}
+            onclick={() => {
+                if (linkingEncounter === null || selectedLinkingSession === null) return;
+                linkEncounterToSession(linkingEncounter, selectedLinkingSession);
+                linkingEncounter = null;
+            }}
+        >
+            Link
+        </Button>
     </div>
-{/if}
+
+</Modal>
+
 
 <style>
     .encounters-page {
@@ -510,65 +471,8 @@ library.add(faLink)
         margin: 0 auto;
     }
 
-    .encounter-form {
-        background: #f8f8f8;
-        padding: 1.5rem;
-        border-radius: 8px;
+    .creator {
         margin-bottom: 2rem;
-    }
-
-    .section {
-        margin: 1.5rem 0;
-        padding: 1rem;
-        background: #fff;
-        border-radius: 4px;
-    }
-
-    .list-item {
-        display: grid;
-        grid-template-columns: minmax(200px, 1fr) auto auto auto auto;
-        gap: 1rem;
-        padding: 0.5rem 1rem;
-        background: #f8f8f8;
-        border-radius: 4px;
-        margin-bottom: 0.5rem;
-        align-items: center;
-    }
-
-    .entity-name {
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-    }
-
-    .entity-xp, .entity-level {
-        white-space: nowrap;
-        color: #666;
-    }
-    .encounters-list {
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-        max-width: 100%;
-    }
-
-    .encounter-card {
-        margin-bottom: 0.5rem;
-        border: 1px solid #e5e7eb;
-        border-radius: 4px;
-        background: white;
-        width: 100%;
-    }
-
-    .encounter-header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        padding: 1rem 1.5rem;
-        cursor: pointer;
-        user-select: none;
-        background: #f9fafb;
-        border-radius: 4px;
     }
 
     .encounter-summary {
@@ -588,15 +492,14 @@ library.add(faLink)
         gap: 2rem;
         align-items: center;
         font-size: 0.875rem;
-        color: #666;
+        color: var(--color-text-secondary);
     }
 
     .encounter-details {
         padding: 1.5rem;
-        border-top: 1px solid #e5e7eb;
-        background: white;
     }
 
+    /* TODO: Get rid of status- or replace with linked */
     .status {
         padding: 0.25rem 0.75rem;
         border-radius: 999px;
@@ -626,22 +529,10 @@ library.add(faLink)
         color: #1f2937; 
     }
 
-    .encounters-section {
-        margin-top: 2rem;
-        background: white;
-        border-radius: 8px;
-        padding: 1.5rem;
-        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-    }
-
-    .encounters-controls {
+    .filter-sort {
         margin-bottom: 1.5rem;
         padding: 1rem;
-        background: #f9fafb;
         border-radius: 4px;
-    }
-
-    .filter-sort {
         display: flex;
         gap: 1rem;
         align-items: center;
@@ -659,23 +550,6 @@ library.add(faLink)
         display: flex;
         gap: 0.5rem;
         align-items: center;
-    }
-
-    .sort-controls select {
-        padding: 0.5rem 1rem;
-        border: 1px solid #e5e7eb;
-        border-radius: 4px;
-        font-size: 0.875rem;
-        min-width: 150px;
-    }
-
-    .sort-direction {
-        padding: 0.5rem 0.75rem;
-        border: 1px solid #e5e7eb;
-        border-radius: 4px;
-        background: white;
-        cursor: pointer;
-        font-size: 1rem;
     }
 
     .detail-section {
@@ -699,7 +573,7 @@ library.add(faLink)
 
     .detail-section li {
         padding: 0.25rem 0;
-        color: #6b7280;
+        color: var(--color-text-secondary);
     }
 
     .actions {
@@ -707,34 +581,7 @@ library.add(faLink)
         gap: 0.75rem;
         margin-top: 1.5rem;
         padding-top: 1rem;
-        border-top: 1px solid #e5e7eb;
-    }
-
-    .modal {
-        position: fixed;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-        background: rgba(0,0,0,0.5);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .modal-content {
-        background: white;
-        padding: 2rem;
-        border-radius: 8px;
-        max-width: 500px;
-        width: 90%;
-    }
-
-    .checkbox-label {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        margin: 0.5rem 0;
+        border-top: 1px solid var(--color-bg-light-raised-border);
     }
 
     .modal-actions {
@@ -744,220 +591,25 @@ library.add(faLink)
         justify-content: flex-end;
     }
 
-    .complete-button {
-        background: #22c55e;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .complete-button:disabled {
-        background: #9ca3af;
-        cursor: not-allowed;
-    }
-
-    .edit-button {
-        background: #3b82f6;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .delete-button {
-        background: #ef4444;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .clone-encounter-button {
-        background: #3b82f6;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .cancel-button {
-        background: #6b7280;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .selected-creatures {
-        margin-bottom: 1rem;
-    }
-
-    .selected-creature {
-        display: flex;
-        align-items: center;
-        gap: 1rem;
-        padding: 0.5rem;
-        background: #f8f8f8;
-        border-radius: 4px;
-        margin-bottom: 0.5rem;
-    }
-
-    .remove-button {
-        margin-left: auto;
-        background: #ef4444;
-        color: white;
-        border: none;
-        padding: 0.25rem 0.5rem;
-        border-radius: 4px;
-        cursor: pointer;
-    }
-
-    .remove-button:hover {
-        background: #dc2626;
-    }
-
-    .list-items {
-        margin-bottom: 1rem;
-    }
-
-    .draft-indicator {
-        position: fixed;
-        bottom: 1rem;
-        right: 1rem;
-        background: white;
-        padding: 0.75rem 1rem;
-        border-radius: 0.5rem;
-        box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        color: #6b7280;
-        font-size: 0.875rem;
-    }
-
-    .draft-badge {
-        background: #3b82f6;
-        color: white;
-        padding: 0.25rem 0.5rem;
-        border-radius: 0.375rem;
-        font-size: 0.75rem;
-        font-weight: 500;
-    }
-
-    .form-header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 1.5rem;
-    }
-
-    .section-header {
-        cursor: pointer;
-        user-select: none;
-    }
-    .toggle-icon {
-        font-size: 0.8em;
-        color: #666;
-    }
-
-    .section-content {
-        padding-top: 1rem;
-    }
-
-    .collapsible {
-        transition: all 0.3s ease;
-    }
-
-    .library-selector-container {
-        margin-top: 1rem;
-        display: flex;
-        gap: 0.5rem;
-    }
-
-    .browse-library-button {
-        display: inline-flex;
-        align-items: center;
-        gap: 0.5rem;
-        background: #3b82f6;
-        color: white;
-        text-decoration: none;
-        padding: 0.1rem 1rem;
-        border-radius: 0.375rem;
-        transition: background-color 0.2s;
-        white-space: nowrap;
-    }
-
-    .browse-library-button:hover {
-        background: #2563eb;
-    }
-
-    .browse-library-button::before {
-        content: "📚";  /* Optional: adds a library emoji */
-    }
-
-    .encounter-form-container {
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-        gap: 1rem;
-    }
-
-    .party-config-row {
-        display: grid;
-        grid-template-columns: 1fr 1fr;
-    }
-
-    .name-input {
-        width: 100%;
-        font-size: 1.2rem;
-        font-family: inherit;
-    }
-
-    .description-input {
-        width: 100%;
-        font-size: 1rem;
-        /* lock size */
-        resize: none;
-        font-family: inherit;
-    }
-
-    .modal select {
-        width: 100%;
-        padding: 0.5rem;
-        margin-bottom: 1rem;
-        border: 1px solid #e5e7eb;
-        border-radius: 4px;
-        font-size: 1rem;
-    }
-
-    .form-group-line {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-    }
 
     .difficulty-trivial {
-        color: #10b981;
+        color: var(--color-difficulty-trivial);
     }
 
     .difficulty-low {
-        color: #f59e0b;
+        color: var(--color-difficulty-low);
     }
 
     .difficulty-moderate {
-        color: #f59e0b;
+        color: var(--color-difficulty-moderate);
     }
 
     .difficulty-severe {
-        color: #ef4444;
+        color: var(--color-difficulty-severe);
     }
 
     .difficulty-extreme {
-        color: #ef4444;
+        color: var(--color-difficulty-extreme);
     }
+
 </style> 
