@@ -4,20 +4,20 @@
 
     interface Props {
         show: boolean;
-        error: string | null;
+        closeButton?: boolean;
+        error?: string;
     }
 
     let { 
         show = $bindable(),
+        closeButton = false,
         error = $bindable(),
-        ...other
     } : Props = $props();
-
     
     const dispatch = createEventDispatcher();
     function closeModal() {
         show = false;
-        error = null;
+        error = undefined;
         dispatch('close');
     }
 </script>
@@ -29,6 +29,13 @@
 
     <div class="modal-backdrop" on:click={closeModal} transition:fade>
         <div class="modal-content" on:click|stopPropagation>
+            <div class="modal-header">
+                <slot name="header" />
+                {#if closeButton}
+                    <button class="close-button" on:click={closeModal}>×</button>
+                {/if}
+            </div>
+
              <slot />
         </div>
     </div>
@@ -49,15 +56,29 @@
         z-index: 1000;
         padding: 2rem;
         overflow-y: auto;
+        
+    }
+
+    .modal-header {
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+    }
+
+    .close-button {
+        background: none;
+        border: none;
+        font-size: 1.5rem;
+        cursor: pointer;
+        color: #666;
     }
 
     .modal-content {
-        background: white;
-        padding: 2rem;
+        background: var(--color-bg);
+        padding: 1.5rem;
         border-radius: 0.5rem;
-        width: 90%;
-        max-width: 1000px;
-        max-height: 90vh;
+        max-width: 95vw;
+        max-height: 95vh;
         overflow-y: auto;
         margin: auto;
     }
