@@ -1,9 +1,9 @@
 <script lang="ts">
     import { createEventDispatcher, onMount } from 'svelte';
     import { fade } from 'svelte/transition';
-    import Modal from './Modal.svelte';
-    import Library from './Library.svelte';
-    import LibraryEntityName from './LibraryEntityName.svelte';
+    import Modal from '../core/Modal.svelte';
+    import Library from '../library/Library.svelte';
+    import LibraryEntityName from '../library/LibraryEntityName.svelte';
     import type { LibraryEntityType } from '$lib/types/library';
     import type { LibraryEntity } from '$lib/types/types';
     import type { CreateOrReplaceEncounter, Encounter } from '$lib/types/encounters';
@@ -11,14 +11,12 @@
     interface Props {
         show: boolean;
         allowedTabs: LibraryEntityType[];
-        addEntityToEncounter?: (entityType : LibraryEntityType, entity : LibraryEntity) => Promise<void>;
         editingEncounter?: CreateOrReplaceEncounter;
     }
 
     let { 
         show = $bindable(),
         allowedTabs,
-        addEntityToEncounter = $bindable(),
         editingEncounter = $bindable()
     } : Props = $props();
     
@@ -33,11 +31,12 @@
 </script>
 
 {#if show}
-    <Modal bind:show={show} on:close={closeModal} bind:error={error}>
-        <Library allowedTabs={allowedTabs} addEntityToEncounter={addEntityToEncounter} bind:editingEncounter={editingEncounter} />
+    <Modal bind:show={show} on:close={closeModal} bind:error={error} closeButton>
+        <div slot="header">
+        </div>
+        <Library allowedTabs={allowedTabs} bind:editingEncounter={editingEncounter} />
     </Modal>
 {/if}
 
 <style>
-
 </style> 
