@@ -3,8 +3,8 @@
   import { onMount } from 'svelte';
   import { browser } from '$app/environment';
   import { auth } from '$lib/stores/auth';
-  import LoginModal from '$lib/components/LoginModal.svelte';
-  import SignupModal from '$lib/components/SignupModal.svelte';
+  import LoginModal from '$lib/components/modals/LoginModal.svelte';
+  import SignupModal from '$lib/components/modals/SignupModal.svelte';
     import { API_URL } from '$lib/config';
 
 </script>
