@@ -16,7 +16,7 @@
     import { API_URL } from '$lib/config';
     import { goto } from '$app/navigation';
     import type { Encounter } from '$lib/types/encounters';
-    import Library from '$lib/components/Library.svelte';
+    import Library from '$lib/components/library/Library.svelte';
 
     export let data: { activeEncounterId: number | null, startTab: string | null };
 
@@ -361,9 +361,7 @@
         <div class="error" transition:fade>{error}</div>
     {/if}
 
-
-    <Library activeEncounterId={data.activeEncounterId} activeTab={activeTab} />
-
+    <Library activeTab={activeTab} />
 </div>
 
 <style>
@@ -373,469 +371,11 @@
         margin: 0 auto;
     }
 
-    .controls {
-        margin-bottom: 2rem;
-    }
-
-    .tabs {
-        display: flex;
-        gap: 0.5rem;
-        margin-bottom: 1rem;
-        border-bottom: 2px solid #e5e7eb;
-        padding-bottom: 0.5rem;
-    }
-
-    .tab-button {
-        padding: 0.5rem 1rem;
-        border: none;
-        background: none;
-        cursor: pointer;
-        font-size: 1rem;
-        color: #6b7280;
-        border-radius: 0.375rem;
-        transition: all 0.2s;
-    }
-
-    .tab-button:hover {
-        background: #f3f4f6;
-        color: #111827;
-    }
-
-    .tab-button.active {
-        background: #3b82f6;
-        color: white;
-    }
-
-    .filters {
-        display: flex;
-        gap: 1rem;
-        margin-bottom: 1rem;
-    }
-
-    .search-input {
-        flex: 1;
-        padding: 0.5rem;
-        border: 1px solid #d1d5db;
-        border-radius: 0.375rem;
-        font-size: 0.875rem;
-    }
-
-    .filter-select {
-        padding: 0.5rem;
-        border: 1px solid #d1d5db;
-        border-radius: 0.375rem;
-        font-size: 0.875rem;
-        min-width: 150px;
-    }
-
-    .table-container {
-        background: white;
-        border-radius: 0.5rem;
-        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-        overflow-x: auto;
-        max-height: 70vh;
-    }
-
-    table {
-        width: 100%;
-        border-collapse: collapse;
-        font-size: 0.875rem;
-    }
-
-    th, td {
-        padding: 0.75rem 1rem;
-        text-align: left;
-        border-bottom: 1px solid #e5e7eb;
-    }
-
-    th {
-        background: #f9fafb;
-        font-weight: 600;
-        color: #374151;
-        position: sticky;
-        top: 0;
-        z-index: 10;
-    }
-
-    th.sortable {
-        cursor: pointer;
-        user-select: none;
-    }
-
-    th.sortable:hover {
-        background: #f3f4f6;
-    }
-
-    .sort-indicator {
-        display: inline-block;
-        margin-left: 0.25rem;
-        transition: transform 0.2s;
-    }
-
-    th.sorted.desc .sort-indicator {
-        transform: rotate(180deg);
-    }
-
-    tr:hover {
-        background: #f9fafb;
-    }
-
     .error {
         background: #fee2e2;
         color: #ef4444;
         padding: 1rem;
         border-radius: 0.375rem;
         margin-bottom: 1rem;
-    }
-
-    .loading {
-        text-align: center;
-        color: #6b7280;
-        padding: 2rem;
-    }
-
-    @media (max-width: 768px) {
-        .filters {
-            flex-direction: column;
-        }
-
-        .filter-select {
-            width: 100%;
-        }
-    }
-
-    .results-count {
-        color: #6b7280;
-        font-size: 0.875rem;
-        margin-top: 0.5rem;
-    }
-
-    .expand-button {
-        background: none;
-        border: none;
-        cursor: pointer;
-        font-size: 1.2rem;
-        color: #6b7280;
-        padding: 0.25rem 0.5rem;
-        border-radius: 0.25rem;
-    }
-
-    .expand-button:hover {
-        background: #f3f4f6;
-    }
-
-    tr.expanded {
-        background: #f8fafc;
-    }
-
-    .detail-row {
-        background: #f8fafc;
-    }
-
-    .entity-details {
-        padding: 1.5rem;
-        background: white;
-        border-radius: 0.5rem;
-    }
-
-    .detail-section {
-        margin-bottom: 1.5rem;
-    }
-
-    .detail-section:last-child {
-        margin-bottom: 0;
-    }
-
-    .detail-grid {
-        display: grid;
-        grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-        gap: 1rem;
-    }
-
-    .detail-item {
-        display: flex;
-        flex-direction: column;
-        gap: 0.25rem;
-    }
-
-    .detail-label {
-        font-weight: 500;
-        color: #4b5563;
-        font-size: 0.875rem;
-    }
-
-    .detail-value {
-        color: #111827;
-    }
-
-    .external-link a {
-        color: #3b82f6;
-        text-decoration: none;
-        display: inline-flex;
-        align-items: center;
-        gap: 0.25rem;
-    }
-
-    .external-link a:hover {
-        text-decoration: underline;
-    }
-
-    .preview-card {
-        position: fixed;
-        background: white;
-        padding: 1rem;
-        border-radius: 0.5rem;
-        box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
-        max-width: 300px;
-        z-index: 50;
-    }
-
-    .preview-card h3 {
-        margin: 0 0 0.5rem 0;
-        color: #111827;
-    }
-
-    .preview-card p {
-        margin: 0;
-        color: #4b5563;
-        font-size: 0.875rem;
-        line-height: 1.4;
-    }
-
-    .actions {
-        display: flex;
-        gap: 0.5rem;
-    }
-
-    .add-button {
-        background: #3b82f6;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 0.375rem;
-        font-size: 0.875rem;
-        cursor: pointer;
-        transition: background-color 0.2s;
-        white-space: nowrap;
-    }
-
-    .add-button:hover {
-        background: #2563eb;
-    }
-
-    .header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 2rem;
-    }
-
-    .mode-indicator-container {
-        display: flex;
-        flex-direction: column;
-        gap: 0.5rem;
-    }
-
-    .mode-indicator {
-        display: flex;
-        align-items: center;
-        gap: 1rem;
-    }
-
-    .mode-badge {
-        background: #3b82f6;
-        color: white;
-        padding: 0.25rem 0.5rem;
-        border-radius: 0.375rem;
-        font-size: 0.875rem;
-        font-weight: 500;
-    }
-
-    .encounter-name {
-        color: #374151;
-        font-weight: 500;
-    }
-
-    .start-draft {
-        background: #3b82f6;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 0.375rem;
-        cursor: pointer;
-        font-size: 0.875rem;
-        transition: background-color 0.2s;
-    }
-
-    .start-draft:hover {
-        background: #2563eb;
-    }
-
-    .exit-mode {
-        background: #6b7280;
-        color: white;
-        border: none;
-        padding: 0.25rem 0.75rem;
-        border-radius: 0.375rem;
-        cursor: pointer;
-        font-size: 0.875rem;
-        transition: background-color 0.2s;
-    }
-
-    .exit-mode:hover {
-        background: #4b5563;
-    }
-
-    .encounter-indicator {
-        position: fixed;
-        bottom: 1rem;
-        right: 1rem;
-        background: white;
-        padding: 0.75rem 1rem;
-        border-radius: 0.5rem;
-        box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-        display: flex;
-        align-items: center;
-        gap: 0.75rem;
-        z-index: 50;
-    }
-
-    .view-encounter {
-        color: #3b82f6;
-        text-decoration: none;
-        font-size: 0.875rem;
-    }
-
-    .view-encounter:hover {
-        text-decoration: underline;
-    }
-
-    .activate-mode-button {
-        background: #3b82f6;
-        color: white;
-        border: none;
-        padding: 0.5rem 1rem;
-        border-radius: 0.375rem;
-        cursor: pointer;
-        font-size: 0.875rem;
-        transition: background-color 0.2s;
-    }
-
-    .activate-mode-button:hover {
-        background: #2563eb;
-    }
-
-    .notification {
-        position: fixed;
-        top: 1rem;
-        right: 1rem;
-        background: #34d399; /* Green background for success */
-        color: white;
-        padding: 0.75rem 1rem;
-        border-radius: 0.5rem;
-        box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-        z-index: 100;
-    }
-
-    .toggle-range-container {
-        margin-top: 0.5rem;
-    }
-
-    .toggle-range-label {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        font-size: 0.875rem;
-        color: #374151;
-    }
-
-    .toggle-range-label input {
-        cursor: pointer;
-    }
-
-    .table-container tr:nth-child(even) {
-        background-color: #f9fafb; /* Light background for even rows */
-    }
-
-    .table-container tr:nth-child(odd) {
-        background-color: white; /* Default background for odd rows */
-    }
-
-    .column-selector-container {
-        margin-bottom: 1rem;
-    }
-
-    .column-selector-toggle {
-        background: #f3f4f6;
-        color: #111827;
-        border: 1px solid #d1d5db;
-        padding: 0.5rem 1rem;
-        border-radius: 0.375rem;
-        cursor: pointer;
-        font-size: 0.875rem;
-        transition: background-color 0.2s;
-    }
-
-    .column-selector-toggle:hover {
-        background: #e5e7eb;
-    }
-
-    .column-selector {
-        background: white;
-        border: 1px solid #d1d5db;
-        border-radius: 0.375rem;
-        padding: 1rem;
-        margin-top: 0.5rem;
-    }
-
-    .column-selector h4 {
-        margin: 0 0 0.75rem 0;
-        color: #374151;
-    }
-
-    .column-options {
-        display: grid;
-        grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-        gap: 0.5rem;
-    }
-
-    .column-option {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        font-size: 0.875rem;
-        color: #374151;
-        cursor: pointer;
-    }
-
-    .column-option input {
-        cursor: pointer;
-    }
-
-    .rarity-label {
-        display: inline-block;
-        padding: 0.25rem 0.5rem;
-        border-radius: 0.375rem;
-        font-size: 0.75rem;
-        font-weight: 500;
-        text-transform: uppercase;
-        letter-spacing: 0.05em;
-        color: white;
-    }
-
-    .rarity-label.common {
-        background-color: #9ca3af; /* Grey */
-    }
-
-    .rarity-label.uncommon {
-        background-color: #fbbf24; /* Yellow */
-    }
-
-    .rarity-label.rare {
-        background-color: #3b82f6; /* Blue */
-    }
-
-    .rarity-label.unique {
-        background-color: #8b5cf6; /* Purple */
     }
 </style> 
